+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Credit card scenarios
+- Delivery with Geolocation scenario
+- Pickup without google maps scenario
+
 ## [0.0.2] - 2019-04-10
 
 ### Added